requirements.txt
gpytranslate
Pyrogram
TgCrypto
pytesseract
Pillow
py-multiapi
tortoise-orm
sqlite-utils
sqlalchemy
toposort

# optional for postgresql
# for db backup

psycopg2-binary

# for tortoise-orm

asyncpg

# optional for mysql

# for db backup

mysqlclient

# for tortoise-orm

aiomysql

# or

asyncmy
